Remove if 0 branching in da_append and da_append_many
Removing unnecessary checking for 0 by creating a scaling calculation on
resize.

Essentially doing n + capacity * scaling factor.

The current init of 256 and scaling factor of 2 may not be optimal,
FOLLY (Facebook standard library) uses a scaling factor for dynamic
arrays of 1.4, which more aligned with their data.
Initial capacity of 64 might be better for general use cases.
Reserved to macros so they can be redefined by user.

The normal scaling factor is kept because setting macros to be
non-implicit convertible (such as fractions) will wreck havoc with
-Wconversion flag, and lots of casts are needed to suppress those
warnings.

There is also the ability to use a C23 compatible macro which can do
casting with typeof operator, which may be good in future.