NoBuild in Kotlin
I got inspired by your nob.h and decided to make one for Kotlin. It backups the last working version and restores it when necessary. Im not packaging a jar or compiling a native executable, only compiling the class files, this makes the whole build time a lot faster. It also uses a Kotlin daemon for having a live compiler environment for faster subsequent builds.
A simple "Hello world" now compiles in 70ms.